Module oak_functions_containers_app::proto::oak::attestation::v1

Modules

attestation_results

Nested message and enum types in AttestationResults.

binary_reference_value

Nested message and enum types in BinaryReferenceValue.

endorsements

Nested message and enum types in Endorsements.

extracted_evidence

Nested message and enum types in ExtractedEvidence.

kernel_binary_reference_value

Nested message and enum types in KernelBinaryReferenceValue.

reference_values

Nested message and enum types in ReferenceValues.

root_layer_data

Nested message and enum types in RootLayerData.

Structs

AmdAttestationReport

Values extracted from an AMD SEV-SNP attestation report.

AmdSevReferenceValues

ApplicationKeys

Keys used by the application to derive encryption session keys and to sign arbitrary data. Each of the certificates contains the final layer’s measurement as additional claims.

ApplicationLayerData

Values extracted from the evidence that represents an application running under the Oak Restricted Kernel.

ApplicationLayerEndorsements

ApplicationLayerReferenceValues

Represents an application running under Oak Restricted Kernel.

AttestationResults

Represents a verification result. Can be extended to return certain measurements and other detail to the client for further processing. Nomenclature follows RFC 9334.

BinaryReferenceValue

CbData

Reserved for future use.

CbEndorsements

CbLayerReferenceValues

CbReferenceValues

CertificateAuthority

Message for passing embedded certificate authority information between layers. Will never appear in the evidence that is sent to the client.

ContainerLayerData

Values extracted from the evidence that represents the Container Runtime Bundle used in Oak Containers.

ContainerLayerEndorsements

ContainerLayerReferenceValues

DiceData

Message that is sent between DICE layers.

Digests

Verifies that the field contains at least one of the given digests. No checks are performed if this is empty. A match in at least one digest is considered a success.

EndorsementReferenceValue

Verifies the transparency log entry, including signatures and the digest.

Endorsements

Endorsement contains statements that some entity (e.g. a hardware provider) vouches for the integrity of claims about the TEE or the software running on it.

Evidence

Attestation Evidence used by the client to the identity of firmware and software running inside a Trusted Execution Environment.

ExtractedEvidence

Evidence values extracted from attestation evidence during verification.

FakeAttestationReport

Values extracted from a fake attestation report when not running in a TEE.

FileReferenceValue

Reference value for a file including its digests.

FirmwareAttachment

The Transparent Release attachment for Oak Stage 0. Measurements are produced with https://github.com/project-oak/oak/snp_measurement.

InsecureReferenceValues

IntelTdxAttestationReport

Values extracted from an Intel TDX attestation report.

IntelTdxReferenceValues

KernelAttachment

The Transparent Release attachment for Oak Containers Linux kernel. Measurements are produced with https://github.com/project-oak/oak/oak_kernel_measurement.

KernelBinaryReferenceValue

Follows the lines of BinaryReferenceValue, but provides a custom proto to facilitate the digest matching.

KernelDigests

Similar to the Digests message, but allows to specify digests for the split components of the bzImage separately.

KernelLayerData

Values extracted from the the kernel layer evidence, as measured by stage0.

KernelLayerEndorsements

KernelLayerReferenceValues

Reference values of the kernel layer, as measured by stage0.

LayerEvidence

DICE layer evidence containing a certificate signed by the previous layer.

OakContainersData

Values extracted from the evidence for an Oak Containers instance.

OakContainersEndorsements

OakContainersReferenceValues

OakRestrictedKernelData

Values extracted from the evidence for a restricted kernel application.

OakRestrictedKernelEndorsements

OakRestrictedKernelReferenceValues

ReferenceValues

RootLayerData

Values extracted from the root layer evidence.

RootLayerEndorsements

RootLayerEvidence

Evidence generated by the Layer0.

RootLayerReferenceValues

SkipVerification

StringReferenceValue

Verifies that a particular string is equal to at least one of the specified ones. No checks are performed if this is empty.

SystemLayerData

Values extracted from the evidence that represents the Oak Containers system image.

SystemLayerEndorsements

SystemLayerReferenceValues

TcbVersion

The versions of the components in the AMD SEV-SNP platform Trusted Compute Base (TCB).

TransparentReleaseEndorsement

Endorsement for binaries from the Transparent Release process. <https://github.com/project-oak/oak/blob/main/docs/release.md>

Enums

TeePlatform

This proto defines the layered DICE Attestation Evidence.