Struct oak_functions_containers_app::proto::oak::functions::oak_functions_client::OakFunctionsClient

pub struct OakFunctionsClient<T> { /* private fields */ }

Implementations

impl OakFunctionsClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error>where D: TryInto<Endpoint>, D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> OakFunctionsClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> OakFunctionsClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn initialize( &mut self, request: impl IntoRequest<InitializeRequest> ) -> Result<Response<InitializeResponse>, Status>

Initializes the service and remote attestation keys.

method_id: 0

pub async fn handle_user_request( &mut self, request: impl IntoRequest<InvokeRequest> ) -> Result<Response<InvokeResponse>, Status>

Handles an invocation coming from a client.

method_id: 1

pub async fn extend_next_lookup_data( &mut self, request: impl IntoRequest<ExtendNextLookupDataRequest> ) -> Result<Response<ExtendNextLookupDataResponse>, Status>

Extends the next lookup data by the given chunk of lookup data. Only after the sender calls finishes building the next lookup data, the receiver replaces the current lookup data with the next lookup data, and only then chunk is will be served in lookups.

method_id: 2

pub async fn finish_next_lookup_data( &mut self, request: impl IntoRequest<FinishNextLookupDataRequest> ) -> Result<Response<FinishNextLookupDataResponse>, Status>

Finishes building the next lookup data with the given chunk of lookup data. The receiver replaces the current lookup data and the next lookup data will be served in lookups.

method_id: 3

pub async fn abort_next_lookup_data( &mut self, request: impl IntoRequest<Empty> ) -> Result<Response<AbortNextLookupDataResponse>, Status>

Aborts building the next lookup data.option

method_id: 4

pub async fn stream_lookup_data( &mut self, request: impl IntoStreamingRequest<Message = LookupDataChunk> ) -> Result<Response<FinishNextLookupDataResponse>, Status>

Streaming version combining ExtendNextLookupData and FinishNextLookupData.

This is mainly for use with gRPC, as microRPC doesn’t support streaming.

method_id: 5

pub async fn reserve( &mut self, request: impl IntoRequest<ReserveRequest> ) -> Result<Response<ReserveResponse>, Status>

Reserves additional capacity for entries in the lookup table.

It should be called before ExtendNextLookupData/StreamLookupData to reduce the number of memory allocations, but it’s not mandatory to call this RPC.

method_id: 6