Trait oak_functions_containers_app::proto::oak::functions::oak_functions_server::OakFunctions

pub trait OakFunctions: Send + Sync + 'static {
    // Required methods
    fn initialize<'life0, 'async_trait>(
        &'life0 self,
        request: Request<InitializeRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<InitializeResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_user_request<'life0, 'async_trait>(
        &'life0 self,
        request: Request<InvokeRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<InvokeResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn extend_next_lookup_data<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ExtendNextLookupDataRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ExtendNextLookupDataResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn finish_next_lookup_data<'life0, 'async_trait>(
        &'life0 self,
        request: Request<FinishNextLookupDataRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FinishNextLookupDataResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn abort_next_lookup_data<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Empty>
    ) -> Pin<Box<dyn Future<Output = Result<Response<AbortNextLookupDataResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn stream_lookup_data<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Streaming<LookupDataChunk>>
    ) -> Pin<Box<dyn Future<Output = Result<Response<FinishNextLookupDataResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn reserve<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ReserveRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ReserveResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with OakFunctionsServer.

Required Methods

fn initialize<'life0, 'async_trait>( &'life0 self, request: Request<InitializeRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<InitializeResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Initializes the service and remote attestation keys.

method_id: 0

fn handle_user_request<'life0, 'async_trait>( &'life0 self, request: Request<InvokeRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<InvokeResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Handles an invocation coming from a client.

method_id: 1

fn extend_next_lookup_data<'life0, 'async_trait>( &'life0 self, request: Request<ExtendNextLookupDataRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<ExtendNextLookupDataResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Extends the next lookup data by the given chunk of lookup data. Only after the sender calls finishes building the next lookup data, the receiver replaces the current lookup data with the next lookup data, and only then chunk is will be served in lookups.

method_id: 2

fn finish_next_lookup_data<'life0, 'async_trait>( &'life0 self, request: Request<FinishNextLookupDataRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<FinishNextLookupDataResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Finishes building the next lookup data with the given chunk of lookup data. The receiver replaces the current lookup data and the next lookup data will be served in lookups.

method_id: 3

fn abort_next_lookup_data<'life0, 'async_trait>( &'life0 self, request: Request<Empty> ) -> Pin<Box<dyn Future<Output = Result<Response<AbortNextLookupDataResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Aborts building the next lookup data.option

method_id: 4

fn stream_lookup_data<'life0, 'async_trait>( &'life0 self, request: Request<Streaming<LookupDataChunk>> ) -> Pin<Box<dyn Future<Output = Result<Response<FinishNextLookupDataResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Streaming version combining ExtendNextLookupData and FinishNextLookupData.

This is mainly for use with gRPC, as microRPC doesn’t support streaming.

method_id: 5

fn reserve<'life0, 'async_trait>( &'life0 self, request: Request<ReserveRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<ReserveResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Reserves additional capacity for entries in the lookup table.

It should be called before ExtendNextLookupData/StreamLookupData to reduce the number of memory allocations, but it’s not mandatory to call this RPC.

method_id: 6